vlm_ui/requirements.txt
streamlit
pillow
requests
fitz
